requirements.txt
securicad-enterprise
networkx
numpy

